@@ -18,7 +18,7 @@ package es.voghdev.chucknorrisjokes.app
 import android.R
 import android.app.Activity
 import android.content.Context
-import android.support.v4.app.Fragment
+import androidx.core.app.Fragment
 import android.view.View
 import android.view.inputmethod.InputMethodManager
 import android.widget.ArrayAdapter
