@@ -16,7 +16,7 @@
 package es.voghdev.chucknorrisjokes.ui.fragment
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.core.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
